service worker: Support DataPipeGetters in navigation request body.
This removes a DCHECK that there are no DataPipeGetters in the body. It
was originally added because I thought it wasn't possible and didn't
want to manually clone the body to handle that case. But it is possible,
and it's no longer necessary to manually clone after r543639
<https://chromium-review.googlesource.com/c/chromium/src/+/963085/>
anyway.

This adds a test that would fail before this CL.

Bug: 944145
Change-Id: Ia7ceba438e741150508cac5fa3d6c56c08137642