Get the preferred main size from a size-contained flex item's box.
Assuming 0 won't do in all cases, at least not for multicol, since we
have to take column-gap and column-width into account. With contain:size
we're supposed to pretend that a box has no contents, which doesn't
imply that we can ignore column-gap and column-width.

This only fixes it when the main axis is the inline axis. When it's the
block axis, it might be safe to return 0 (which is what we still do).

This improves the rendering of
css/vendor-imports/mozilla/mozilla-central-reftests/contain/contain-size-multicol-003.html
but we still fail.

Bug: 863454
Change-Id: I444f94321dc1b10741515d8527ed4fdd72e6a15e