[ElementTiming] Add background image support
This CL adds background image support to Element Timing. For each
element and each background image affecting it, we have a potential new
entry. Therefore, in order to keep track of the first paint, we keep a
HashSet of pairs of pointers (LayoutObject, ImageResourceContent). We
need to remove a pair from the HashSet when the corresponding
LayoutObject is no longer observing the image, so we add a new method in
ImageResourceObserver to support this removal.

Bug: 879270
Change-Id: If7bc70df30ba0a9ed8335e67e09f52d6f9ba849d